Refactor the decompiler class
