+++ thoth/thoth.py
@@ -146,7 +146,7 @@ def main() -> int:
         a = analyzer(disassembler, color=args.color)
         if a._detect():
             detected = True
-            detected_analyzers_count += 1 
+            detected_analyzers_count += 1
         a._print()
         if detected:
             print()
@@ -159,7 +159,7 @@ def main() -> int:
             selected_analyzers_count,
             "s" if selected_analyzers_count > 1 else "",
             "were" if selected_analyzers_count > 1 else "was",
-            detected_analyzers_count
+            detected_analyzers_count,
         )
     )
     return 0
